refactor: move circuit breaker utility to shared package
- Create generic circuit breaker in shared with logger interface
- Move circuit breaker logic to packages/shared/src/utils/circuit-breaker.ts
- Update proxy circuit-breaker.ts to use shared implementation with logger adapter
- Export circuit breaker utilities from shared package index
- Maintain backward compatibility for proxy service
- Fix unused variable issue
